pyproject.toml
# ---------------------------------------------------------------------------
#  PYPROJECT.TOML  ·  Price-Transparency Query Engine
#  Compatible with Poetry ≥1.5 (PEP-621 metadata)
# ---------------------------------------------------------------------------

[project]
name = "price-transparency-engine"
version = "0.1.0"
description = "Price-Transparency Query Engine"
readme = "README.md"
authors = [
  { name = "vin72", email = "49925973+vin72@users.noreply.github.com" }
]

requires-python = ">=3.11,<4.0"

dependencies = [
  "snowflake-connector-python>=3.15.0,<4.0.0",
  "fastapi>=0.115.12,<0.116.0",
  "uvicorn[standard]>=0.34.2,<0.35.0",
  "pydantic>=2.11.3,<3.0.0",
  "langchain>=0.3.24,<0.4.0",
  "streamlit>=1.44.1,<2.0.0",
  "python-dotenv>=1.1.0,<2.0.0",
  "pyarrow>=18.0.0,<19.0.0",        # ← pinned for Snowflake pandas helpers
  "duckdb (>=1.2.2,<2.0.0)",
  "pandas (>=2.2.3,<3.0.0)"
]

[project.scripts]
snowflake-shell = "src.db:cli"      # →  poetry run snowflake-shell
cms-refresh = "src.ingest.ingest_cms:ingest"
# ---------------------------------------------------------------------------
#  Poetry-specific configuration
# ---------------------------------------------------------------------------

[tool.poetry]
packages = [{ include = "src" }]    # tell Poetry to install code in src/

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

# ---------------------------------------------------------------------------
#  Build backend (standard for Poetry projects)
# ---------------------------------------------------------------------------

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
